backend/requirements.txt
django==2.2.19
gunicorn==20.0.4
psycopg2-binary==2.9.0
python-dotenv==0.20.0
djangorestframework==3.13.1
Pillow==9.1.0
django-filter==0.4.2
